Update 45_Popularity.asciidoc
If I am understanding example and how the "missing" property works, then I think my suggestion helps to make more people aware of their development options.  In this example, I think the value of the "missing" property would have to be strictly between 0 and 1 in order to get the desired boosting.