Fixed error in Who Should Read This Book section
The sentence " We expect you to have some programming background and, although not required, it would help
to have used SQL and a relational database." looks a bit confusing. Proposing to change it to
" We expect you to have some programming background and, although not required, it would help
to have used SQL and a relational database knowledge." instead.