Update 35_Tutorial_Aggregations.asciidoc
The example queries result in this error:

    Fielddata is disabled on text fields by default. Set fielddata=true on [interests] in order to load fielddata in memory by uninverting the inverted index. Note that this can however use significant memory.

Don't really understand what's going on yet as this is my first day with ES, but I found using `"field": "interests.keyword"` instead of `"field": "interests"` gives the desired result. It'd be good if an easy to understand explanation for the need of this `.keyword` was provided too (people at this page are mostly beginners).