Filtering Queries and Aggregations result improvement
```
We want to display search results according to what the user searches for. But we also want to enrich the page by including the average price of cars (matching the search) that were sold in the last month.
```

Therefore there should be search results in the example. This confused me because I'm trying to understand what the code is doing, but it doesn't line up with the description above.